Tests to explore the much faster HIF device
Run as follows:
./ehwe -d i2c:1:hif:master:0 -- -tn

Where 'n' is the number of the test. n=0 is the old test.

The new HIF device under development is based on different
technology than BusPirate. It's 100x faster, potentially much
faster. I.e. it transfers many times faster than a PWM cycle.

New tests explores what happens when several commands occur
within one PWM-cycle.

This also happens to be a compliance-test as it's the first
real-life case where HIF is used. I.e. validation of that
work-bench code is fully transparent to the bus-interface.

Change-Id: Ia219fda495e2c7ebc053b71df16f0862f058e3be