Temperature correctly weighted from raw ADC
* Transform to true 10-bit ADC representation without reading
  i2c twice, which would read inconsistent values occasionally
  (y2k)
* 10-bit value is to temperature linear but 4 bits shift-left to
  physical temperature. Transform to float and divide in one go.

NOTE: ntohs(..) is a not correct as integer representations are
device to device oriented. Network bit- and byte-order
transformation isn't what's desired but happens to work on x86 &
(default) ARM.  Until support for LSB/MSB (i.e. NOT bit-order)
support is added into the high level (i.e. register access)
i2c-driver, this is used as a workaround.